lesson 20.

+++ laravel-qa/app/Http/Controllers/QuestionsController.php
@@ -76,6 +76,11 @@ class QuestionsController extends Controller
     // public function edit($id)
     {
         // $question = Question::findOrFail($id);
+
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'Access denied!');
+        }
+
         return view('questions.edit', compact('question'));
     }
 
@@ -88,6 +93,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'Access denied!');
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect(route('questions.index'))->with('success', 'Your question has been updated!');
@@ -101,6 +110,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, 'Access denied!');
+        }
+
         $question->delete();
 
         return redirect(route('questions.index'))->with('success', 'Your question has been deleted!');
